enclosing em-$$ in closing double quote
moved closing double quote mark to include the double dollar sign.  In OpenRC, having the dollar sign outside of the quotation mark causes chroot.py to not find the temporary pid file.  (Not fully tested, e.g. going back to original form and generating error and then making change and seeing that the step/task "locale" works as I  do not know where the log ares cached and I've asked in Issue #179 and on Discord general chat how to force that particular step to be reprocessed.